+++ gui/screens/credits_screen.gd
@@ -1,0 +1,17 @@
+# Open Street Kart is an arcade kart game where you race in real life areas reconstructed from Open Street Map
+# Copyright (c) 2025 Charly Schmidt aka Picorims<picorims.contact@gmail.com> and Open Street Kart contributors
+
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+class_name CreditsScreen extends Control
+
+signal back_requested
+
+func _ready() -> void:
+	var back_button = $BackButton
+	assert(back_button != null, "back button undefined.")
+	back_button.pressed.connect(func():
+		back_requested.emit()
+	)